rename module to `codec`
