+++ packages/svelte-docgen/src/codec.js
