+++ packages/svelte-docgen/src/mod.js
@@ -3,4 +3,4 @@ export { analyzeComponent } from "./analyzer/component.js";
 export { analyzeProperty } from "./analyzer/prop.js";
 export { Options } from "./options.js";
 export { parse } from "./parser/mod.js";
-export { deserialize, serialize } from "./serde.js";
+export { deserialize, serialize } from "./codec.js";
